First experiment with a stack-machine
Here we set the scene for how a stack machine can work. We simulate a
simple, hard-coded program which just puts zero onto the stack, then
increments it. At the end, we print the stack.

Try experimenting by customising the hard-coded program, and see what
you end up with.

Can you add an instruction to take the top two numbers from the stack,
add them together and put it back on the stack?