Add scaffolding files
These files aren't code, but are useful for the tools we use to edit and
communicate code.

The `.editorconfig` is used to tell text editors some basic information
about styling the file, so that it is more consistent between
developers.

The `.gitignore` tells git about some files/patterns that we don't want
it to track, for example output generated by the compiler. We don't want
to track this because it is different from machine to machine, and the
inputs to the compiler are tracked by git.